convert mispec tests to NTest and remove mispec
